feat: adds `replSource` property and command-line option
The `replSource` property specifies the frege source file that you want to
load into the repl. It is excluded in the `fregeCompile` task so that
we don't get two java class files (one from `compileFrege` and one
from the fregeRepl `:l` command) that shadow each other on the classpath.

As a result, we can make interactive changes to the `replSource` file and
use the `:r` reload command to see them.