feat: simplifies the replFrege task
The replFregeTask has the following new logic:
1. Compiles the specified fregeRepl module (either in the
   `build.gradle` via command line option `--replModule=...`) and all
   its dependencies.
2. Sets up the correct classpath so that dependent modules don't have
to be imported manually. In addition, it solves the shadowing problem by
removing the replModule java and class file from the classpath.
3. It prints one single command to directly start the repl and load the
   specified module.

Bonus: I designed the task so that you can even automate step 3 with
the following bash command: `eval $(./gradlew -q replFrege)`.